updated code w/ partners

+++ src/ResourceControl.java
@@ -1,15 +1,15 @@
 public class ResourceControl implements ResourceOptimization {
-    public void optimizeWaterUsage(int irrigationDuration, int irrigationRate, int waterUsed, int numberOfIrrigationSessions)
+    public void optimizeWaterUsage(int irrigationDuration, int irrigationRate, double waterUsed, int numberOfIrrigationSessions)
     {
         int waterUsePerSession = irrigationDuration * irrigationRate;
         System.out.println("Water used per session is " + waterUsePerSession);
 
-        int totalWaterUsage = waterUsed * numberOfIrrigationSessions;
+        double totalWaterUsage = waterUsed * numberOfIrrigationSessions;
         System.out.println("Total water used per day is " + totalWaterUsage);
     }
-    public void optimizeEnergyUsage(int powerConsumption, int irrigationDuration)
+    public void optimizeEnergyUsage(double powerConsumption, double irrigationDuration)
     {
-        int totalEnergyUsage = powerConsumption * irrigationDuration;
+        double totalEnergyUsage = powerConsumption * irrigationDuration;
         System.out.println("Total energy usage per day is " + totalEnergyUsage);
     }
 }
